Working process branch

+++ src/view/BudgetTab.java
@@ -14,7 +14,6 @@ import java.time.LocalDate;
 import java.util.HashMap;
 import java.text.DecimalFormat;
 import model.Document;
-import model.Project;
 
 /**
  * The budget tab for a project in the FileNtro program.
@@ -84,6 +83,7 @@ public class BudgetTab extends JPanel {
      *
      * @param theProjectID The ID of the project
      * @author Thinh Le
+     * @author Riley Bennett
      */
     public BudgetTab(final String theProjectID) {
 
@@ -91,6 +91,7 @@ public class BudgetTab extends JPanel {
 
         // Get documents associated with the project.
         myDoc = DocumentController.getDocsByProjectID(theProjectID);
+        totalBudget = ProjectController.getTotalBudgetByID(theProjectID).doubleValue();
 
         setLayout(new GridBagLayout());
 
@@ -124,20 +125,11 @@ public class BudgetTab extends JPanel {
         GridBagConstraints c = new GridBagConstraints();
         c.fill = GridBagConstraints.HORIZONTAL;
         c.insets = new Insets(15, 15, 15, 15);
-        totalBudgetLabel = new JLabel("Total Budget:$");
+        totalBudgetLabel = new JLabel("Total Budget: $" + totalBudget);
+        totalBudgetLabel.setFont(new Font("Verdana", Font.PLAIN, 18));
 
         add(totalBudgetLabel, c);
 
-        // Add budget field
-        setBudgetField = new JTextField();
-        setBudgetField.setText("");
-        setBudgetField.setEditable(false);
-        c.gridx = 1;
-        c.gridy = 0;
-
-        setBudgetField.setColumns(5);
-
-        add(setBudgetField, c);
         c.gridx = 2;
         c.gridy = 2;
 
@@ -231,11 +223,11 @@ public class BudgetTab extends JPanel {
 
                     updateTable();
 
-                    //This is a warning message when it is over your budget.
-                    if(theTotalCost>totalBudget){
+                    // //This is a warning message when it is over your budget.
+                    // if(theTotalCost>totalBudget){
 
-                        JOptionPane.showMessageDialog(BudgetTab.this,"You are over your budget!","Reminder",JOptionPane.WARNING_MESSAGE);
-                    }
+                    //     JOptionPane.showMessageDialog(BudgetTab.this,"You are over your budget!","Reminder",JOptionPane.WARNING_MESSAGE);
+                    // }
                 }
 
 
@@ -271,6 +263,7 @@ public class BudgetTab extends JPanel {
 
                         ProjectController.setTotalBudgetByID(theProjectID, BigDecimal.valueOf(newBudget));
                         setTotalBudget(newBudget);
+                        updateTable();
                     }
 
 
@@ -287,19 +280,23 @@ public class BudgetTab extends JPanel {
         deleteButton.addActionListener(e -> {
             int row = table.getSelectedRow();
             if (row >= 0) {
+                int result = JOptionPane.showConfirmDialog(BudgetTab.this, "Are you sure you want to delete this item?",
+                "Confirmation", JOptionPane.YES_NO_OPTION);
 
-//                System.out.println("Deleting " + name);
-                String id = (String) table.getValueAt(row, 0);
-                theTotalCost -= myDoc.get(id).getTotalCost().doubleValue();
+                if (result == JOptionPane.YES_OPTION) {
+//                  System.out.println("Deleting " + name);
+                    String id = (String) table.getValueAt(row, 0);
+                    theTotalCost -= myDoc.get(id).getTotalCost().doubleValue();
 
-                // Round to 2 decimal places
-                theTotalCost = Double.valueOf(df.format(theTotalCost));
-                totalLabel.setText("Current Cost: $" + theTotalCost); 
-                DocumentController.deleteADocument(myDoc.get(id));
+                    // Round to 2 decimal places
+                    theTotalCost = Double.valueOf(df.format(theTotalCost));
+                    totalLabel.setText("Current Cost: $" + theTotalCost); 
+                    DocumentController.deleteADocument(myDoc.get(id));
 
-                myDoc.remove(id);
+                    myDoc.remove(id);
 
-                updateTable();
+                    updateTable();
+                }
             }
         });
 
@@ -388,7 +385,7 @@ public class BudgetTab extends JPanel {
 
     /**
      * Adds a row to the table.
-     *
+     * @author Thinh Le
      * @param id The ID of the item to be added.
      * @param name The name of the item to be added.
      * @param totalCost The cost of the item to be added.
@@ -401,6 +398,8 @@ public class BudgetTab extends JPanel {
 
     /**
      * Updates the table with new data.
+     * @author Thinh Le
+     * @author Riley Bennett
      */
     public void updateTable() {
         // Clear the rows
@@ -410,14 +409,16 @@ public class BudgetTab extends JPanel {
         myDoc.forEach((k, e) -> {
             addRowToTable(k, e.getDocumentName(), df.format(e.getTotalCost()));
         });
+
         BigDecimal pTotalCost= ProjectController.updateTotalCostByID(this.theProjectID).setScale(2, RoundingMode.CEILING);
         theTotalCost = Double.valueOf(pTotalCost.doubleValue());
-        totalLabel.setText("Current Cost: $" + theTotalCost);
-
+        
+        setCostLabel();
     }
 
     /**
      * Sets the total budget of the project.
+     * @author Thinh Le
      * @param budget The budget to set to.
      */
     public void setTotalBudget(double budget) {
@@ -427,13 +428,16 @@ public class BudgetTab extends JPanel {
 
     /**
      * Updates the label of the total budget.
+     * @author Thinh Le
      */
     public void updateTotalBudgetLabel() {
-        setBudgetField.setText("" + ProjectController.findProjectByID(theProjectID).getBudget());
+        totalBudgetLabel.setText("Total Budget: $" + totalBudget);
     }
 
     /**
      * Updates the total cost of this project.
+     * @author Thinh Le
+     * @author Riley Bennett
      */
     public void updateTotalCost() {
         myDoc.forEach((k, b) -> {
@@ -442,8 +446,21 @@ public class BudgetTab extends JPanel {
 
         // Round to 2 decimal places
         theTotalCost = Double.valueOf(df.format(theTotalCost));
+        setCostLabel();
+    }
 
-        totalLabel.setText("Current Cost: $" + String.format("%.2f", theTotalCost));
+    /**
+     * Sets the cost label to red if the current cost is over budget, or black if cost is less than budget.
+     * @author Riley Bennett
+     */
+    public void setCostLabel() {
+        if (theTotalCost > totalBudget) {
+            totalLabel.setText("Current Cost (Over budget!): $" + theTotalCost);
+            totalLabel.setForeground(Color.RED);
+        } else {
+            totalLabel.setText("Current Cost: $" + theTotalCost);
+            totalLabel.setForeground(Color.BLACK);
+        }
     }
 
     /**
@@ -609,7 +626,6 @@ public class BudgetTab extends JPanel {
 
                                 // Round to 2 decimal places
                                 theTotalCost = Double.valueOf(df.format(theTotalCost));
-                                totalLabel.setText("Current Cost: $" + theTotalCost);
 
                                 updateTable();
                             }
